fastboot: install kpartx
Provisioning a partitioned LUKS container via IDP requires device
mapper nodes to be available. Issuing BLKRRPART on an opened LUKS block
device is insufficient to trigger the creation of these, so install
kpartx which can be invoked to take care of it. kpartx has a slightly
smaller dep chain compared to partprobe and adds minimal overhead.